- Use featureInfoClick action to trigger identify if featureinfo parameter is provided. Supporting limiting of queries layers by passing extra property into parameter value. - Moved "updatePointWithGeometricFilter" into IdentifyUtils
