+++ web/client/epics/identify.js
@@ -45,13 +45,12 @@ import { modeSelector, getAttributeFilters, isFeatureGridOpen } from '../selecto
 import { spatialFieldSelector } from '../selectors/queryform';
 import { mapSelector, projectionDefsSelector, projectionSelector, isMouseMoveIdentifyActiveSelector } from '../selectors/map';
 import { boundingMapRectSelector } from '../selectors/maplayout';
-import { centerToVisibleArea, isInsideVisibleArea, isPointInsideExtent, reproject, reprojectBbox, calculateCircleCoordinates,
-    calculateCircleRadiusFromPixel } from '../utils/CoordinatesUtils';
+import { centerToVisibleArea, isInsideVisibleArea, isPointInsideExtent, reprojectBbox} from '../utils/CoordinatesUtils';
 import { floatingIdentifyDelaySelector } from '../selectors/localConfig';
 import { createControlEnabledSelector, measureSelector } from '../selectors/controls';
 import { localizedLayerStylesEnvSelector } from '../selectors/localizedLayerStyles';
 import { mouseOutSelector } from '../selectors/mousePosition';
-import {getBbox, getCurrentResolution, parseLayoutValue, getHook, GET_COORDINATES_FROM_PIXEL_HOOK, GET_PIXEL_FROM_COORDINATES_HOOK} from '../utils/MapUtils';
+import {getBbox, getCurrentResolution, parseLayoutValue} from '../utils/MapUtils';
 import {buildIdentifyRequest, filterRequestParams} from '../utils/MapInfoUtils';
 import { IDENTIFY_POPUP } from '../components/map/popups';
 
@@ -61,58 +60,8 @@ const gridGeometryQuickFilter = state => get(find(getAttributeFilters(state), f 
 const stopFeatureInfo = state => stopGetFeatureInfoSelector(state) || isFeatureGridOpen(state) && (gridEditingSelector(state) || gridGeometryQuickFilter(state));
 
 import {getFeatureInfo} from '../api/identify';
-import { MAP_TYPE_CHANGED } from './../actions/maptype';
-
-/**
- * Recalculates pixel and geometric filter to allow also GFI emulation for WFS.
- * This information is used also to switch to edit mode (feature grid) from GFI applying the same filter
- * @param {object} point the point clicked, emitted by featureInfoClick action
- * @param {string} projection map projection
- */
-const updatePointWithGeometricFilter = (point, projection) => {
-    // calculate a query for edit
-    const lng = get(point, 'latlng.lng');
-    const lat = get(point, 'latlng.lat');
-    // update pixel if changed
-    const pos = reproject([lng, lat], 'EPSG:4326', projection);
-    const getPixel = getHook(GET_PIXEL_FROM_COORDINATES_HOOK);
-    let pixel;
-    if (getPixel) {
-        const [x, y] = getPixel([pos.x, pos.y]);
-        pixel = { x, y };
-    } else {
-        pixel = point.pixel;
-    }
-    const hook = getHook(GET_COORDINATES_FROM_PIXEL_HOOK);
-    const radius = calculateCircleRadiusFromPixel(
-        hook,
-        pixel,
-        pos,
-        5
-    );
-    // emulation of feature info filter to query WFS services (edit and/or WFS layer)
-    const geometricFilter = {
-        type: 'geometry',
-        enabled: true,
-        value: {
-            geometry: {
-                center: [pos.x, pos.y],
-                coordinates: calculateCircleCoordinates(pos, radius, 12),
-                extent: [pos.x - radius, pos.y - radius, pos.x + radius, pos.y + radius],
-                projection,
-                radius,
-                type: "Polygon"
-            },
-            method: "Circle",
-            operation: "INTERSECTS"
-        }
-    };
-    return {
-        ...point,
-        pixel,
-        geometricFilter
-    };
-};
+import { MAP_TYPE_CHANGED } from '../actions/maptype';
+import {updatePointWithGeometricFilter} from "../utils/IdentifyUtils";
 
 /**
  * Epics for Identify and map info
